requirements.txt
annotated-types==0.7.0
anyio==4.6.0
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.1
databases==0.9.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.115.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==8.3.3
python-bcrypt==0.3.2
python-multipart==0.0.9
requests==2.32.3
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
